update xCode schema to solve Bitrise problems
