--- ChatApp/Core/UseCases/UserLogin/LoginParameter.swift
@@ -1,13 +1,0 @@
-//
-//  LoginParamater.swift
-//  ChatApp
-//
-//  Created by mac on 29/11/2022.
-//
-
-import Foundation
-
-struct LoginParameter {
-    let password: String
-    let email: String
-}
