+++ MyPodcasts/Core/UseCases/UserLogin/LoginParameter.swift
@@ -1,0 +1,13 @@
+//
+//  LoginParamater.swift
+//  MyPodcasts
+//
+//  Created by mac on 29/11/2022.
+//
+
+import Foundation
+
+struct LoginParameter {
+    let password: String
+    let email: String
+}
