+++ MyPodcasts/EntityGateway/API/Client/AlamofireNetwork/Network/Alamofire.swift
@@ -1,0 +1,54 @@
+//
+//  Alomafire.swift
+//  MyPodcasts
+//
+//  Created by mac on 15/12/2022.
+//
+
+import Alamofire
+import Combine
+import Foundation
+
+/// AlamofireWrapper: Encapsulates all of the Alamofire OP's
+///
+public class AlamofireNetwork: Network {
+
+    public init() {}
+
+    /// Executes the specified Network Request. Upon completion, the payload will be sent back to the caller as a Data instance.
+    ///
+    public func responseData(for request: URLRequestConvertible,
+                             completion: @escaping (Result<Data, Error>) -> Void) {
+        AF.request(request).responseData { response in
+            completion(response.result.toSwiftResult())
+        }
+    }
+
+    /// Executes the specified Network Request. Upon completion, the payload or error will be emitted to the publisher.
+    /// Only one value will be emitted and the request cannot be retried.
+    ///
+    public func responseDataPublisher(for request: URLRequestConvertible) -> AnyPublisher<Result<Data, Error>, Never> {
+        return Future { promise in
+            AF.request(request).responseData { response in
+                let result = response.result.toSwiftResult()
+                promise(.success(result))
+            }
+        }.eraseToAnyPublisher()
+    }
+}
+
+// MARK: - Swift.Result Conversion
+
+//
+private extension Result where Failure == AFError {
+    /// Convert this `Alamofire.Result` to a `Swift.Result`.
+    ///
+    func toSwiftResult() -> Swift.Result<Success, Error> {
+        switch self {
+        case let .success(value):
+            return .success(value)
+        case let .failure(error):
+            return .failure(error)
+        }
+    }
+}
