+++ MyPodcasts/EntityGateway/API/Client/AlamofireNetwork/Requests/FakeStoreRequest.swift
@@ -1,0 +1,63 @@
+//
+//  File.swift
+//
+//
+//  Created by mac on 15/12/2022.
+//
+
+import Alamofire
+import Foundation
+
+/// Represents Fakestore.com Endpoint
+///
+struct FakeStoreRequest: URLRequestConvertible {
+
+    /// HTTP Request Method
+    ///
+    let method: HTTPMethod
+
+    /// URL Path
+    ///
+    let path: String
+
+    /// Parameters
+    ///
+    let parameters: [String: Any]
+
+    /// Designated Initializer.
+    ///
+    /// - Parameters:
+    ///     - method: HTTP Method we should use.
+    ///     - path: RPC that should be called.
+    ///     - parameters: Collection of Key/Value parameters, to be forwarded to the Jetpack Connected site.
+    ///
+    init(method: HTTPMethod, path: String, parameters: [String: Any]? = nil) {
+        self.method = method
+        self.path = path
+        self.parameters = parameters ?? [:]
+    }
+
+    /// Returns a URLRequest instance reprensenting the current FakeStore Request.
+    ///
+    func asURLRequest() throws -> URLRequest {
+        guard let url = URL(string: Settings.storeApiBaseURL + path) else {
+            throw RequestError.invalidURl
+        }
+        
+        let request = try URLRequest(url: url, method: method, headers: nil)
+
+        return try encoder.encode(request, with: parameters)
+    }
+}
+
+// MARK: - FakeStore Request: Internal
+
+//
+private extension FakeStoreRequest {
+
+    /// Returns the Parameters Encoder
+    ///
+    var encoder: ParameterEncoding {
+        return method == .get ? URLEncoding.queryString : URLEncoding.httpBody
+    }
+}
