+++ MyPodcasts/EntityGateway/API/Client/AlamofireNetwork/Requests/RequestError.swift
@@ -1,0 +1,14 @@
+//
+//  RequestError.swift
+//  MyPodcasts
+//
+//  Created by mac on 15/12/2022.
+//
+
+import Foundation
+
+public enum RequestError: Error {
+    /// invalid url for request!
+    ///
+    case invalidURl
+}
