+++ MyPodcasts/EntityGateway/API/Client/Settings.swift
@@ -1,0 +1,17 @@
+//
+//  File.swift
+//
+//
+// Created by mac on 15/12/2022.
+//
+
+import Foundation
+
+/// Networking Settings
+///
+enum Settings {
+
+    /// Base API URL
+    ///
+    static let storeApiBaseURL = "https://fakestoreapi.com/"
+}
