+++ MyPodcasts/Scenes/Authentication/Onboarding/OnboardingConfigurator.swift
@@ -1,0 +1,19 @@
+//
+//  OnboardingConfigrator.swift
+//  MyPodcasts
+//
+//  Created by mac on 04/12/2022.
+//
+
+import Foundation
+
+// DI Layer
+class OnboardingConfigurator {
+    static func configure() -> OnboardingViewController {
+        let viewController = OnboardingViewController()
+        let presenter = OnboardingPresenterImpl(view: viewController)
+
+        viewController.presenter = presenter
+        return viewController
+    }
+}
